feat(date): changed date offset in order to support more dates
An error due to the need of more bits to store dates, mainly to fix the age of some users, made the
necessity of changing the date offset. Now, instead of subtract the date value to the System Date
Offset, a sum is performed, making a solution to the problem, without needing to increase the memory
usage.