feat(information_user): changed user variable total_spent to information
== INFORMATION
- Now with future bit shifts the information variable from user struct contains values relative to number of reservations
and total spent by a user. This implementation features a memory improvement, since instead using two int variables
in one variable we store two values. Of course, for this dataset this strategy works perfectly, but for a more
professional project, a usage of two variables would be the best aproach.