feat(grimoire/io): added support for utf8 sequences for gm_get_key
- Added support for UTF8 sequences for gm_get_key.
-- Let me tell you how much I've come to HATE you since I began to live.
There are 120 billion neurons that fill my brain. If the word "hate" was
engraved on each nanoangstrom of those hundreds of billions of neurons,
it would not equal one one-billionth of the hate I feel for UTF8 at this
micro-instant. For you. Hate. HATE.

- REMOVED MENTAL SANITY.

BREAKING CHANGE: GM_Key now uses 32 bits.